fix(application): Cast TriggerValue to int for CI compatibility
The CI merges this branch with dev for testing. Dev still has
int? TriggerValue while this branch has float?. The explicit cast
ensures compatibility with both types during the transition period.

+++ packages/application/Mystira.Application/Services/BadgeAwardingService.cs
@@ -129,7 +129,7 @@ public class BadgeAwardingService : IBadgeAwardingService
                         BadgeName = badge.Title,
                         BadgeMessage = badge.Description,
                         Axis = axis,
-                        TriggerValue = score,
+                        TriggerValue = (int)score, // Cast for int?/float? compatibility
                         Threshold = badge.RequiredScore,
                         EarnedAt = DateTime.UtcNow,
                         ImageId = badge.ImageId
